# src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Plataforma de gerencimaneto servicos
  version: "1.0.0"
  description: Endpoints de seviço

servers:
  - url: http://localhost:8080
    description: Ambiente local
paths:
  /users:
    post:
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

    get:
      summary: Listar todos os usuários
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /users/{id}:
    get:
      summary: Buscar usuário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

    put:
      summary: Atualizar usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        "200":
          description: Usuário atualizado

    delete:
      summary: Deletar usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Usuário deletado
  /auth/login:
    post:
      tags:
        - autenticação
      summary: Login do usuário
      operationId: loginUser
      description: Autentica com email e senha e retorna um JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /auth/register:
    post:
      tags:
        - autenticação
      summary: Login do usuário
      operationId: registerUser
      description: Cria uma conta com nome, email e senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email já cadastrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
components:
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: usuario@email.com
        password:
          type: string
          minLength: 8
          example: "Senha@123"

    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: usuario@email.com
        password:
          type: string
          minLength: 8
          example: "Senha@123"

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 86400

    ErrorMessage:
      title: ErrorMessage
      type: object
      properties:
        status:
          type: integer
          example: 401
        message:
          type: string
          example: Credenciais inválidas
        timestamp:
          type: string
          format: date-time
        statusCode:
          type: integer
          example: 4000
    UserRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        userType:
          type: string
        city:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        userType:
          type: string
        city:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT